move some helping funcs
