+++ bank/src/helps.hpp
@@ -8,6 +8,8 @@ using namespace std;
 namespace helps {
     string get_ascii_art();
     filesystem::path find_appdata_path();
+    bool account_exist(string);
+    string generate_random_account_no();
 }
 
 #endif
